+++ Engine/Source/GUI/widget_tree.cpp
@@ -95,7 +95,7 @@ namespace Ming3D
             unsigned int* indData = mIndexData->GetData();
             visual->GetMeshData(&vertData[mVertexIndex], &indData[mTriangleIndex]);
 
-            for (int i = 0; i < indCount; ++i)
+            for (size_t i = 0; i < indCount; ++i)
                 indData[mTriangleIndex + i] += mVertexIndex;
 
             // Add render batch
